third_party/WebKit/LayoutTests/svg/hixie/error/010-expected.txt

CONSOLE ERROR: line 1: Error: <svg> attribute viewBox: Trailing garbage, "0 0 1 1 1".
